[ci] add minimal tests for docker images
closed #83

Add a minimal test suite that validates that the images we build and
publish have a minimally working Theia application, inside.

Until now, the status of the build of an image, determines if we publish
it or not. However an image can carry a non-functional Theia
application even if builds without error. In that case we do not want to
publish it.

The UI test suite and related configurations used were borrowed from
the main theia repo and tweaked. ATM we use only the most basic test
suite, "main-page", that should be applicable to all Theia applications
we publish.

It's possible to add image-specific test files. For example, see
wdio.theia-docker.conf.js, that adds an extra test file to be
run, for the theia-docker image (for now just a second copy of
the "main-page" test file).