[cpp] Added cpp Docker application
The image bundles the nightly build for clangd (C/C++ Language Server)
and command-line clang-tidy, that's suitable to run linting on a
workspace. The preference that enables the clang-tidy integration,
where the one that runs co-located with clangd is used to lint opened
files, and create problem markers in the IDE, about linter issues found.

Fixes #194

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>
Signed-off-by: Dinne Bosman <dwjbosman@yahoo.com>